15 have interest rate compound daily (#38)
* Fix input
Code cleanup

* Start of adding functionality

* Balance code is working.  Need to update tests

* 5

* Doing stuff

* Wrap up unit tests around Contexts

* Remove babel

* asdf

* Always delete transactions regardless of test outcome

* Tests passing and auth route working as expected

* Updates.  Going to try million js

* Added cache invalidation to route for getting child accounts

* Things

* Better define preview components and added a loader overlay

* Seems like core functionality is working.  Time to update/write playwright tests to cover the hiccups with unit tests.

* Added unit test around balance so that we can be sure this is working properly
Updated playwright tests to be better
Fixed some nullable issues around the ChildPreview

+++ tests/e2e/parent/dashboard/requests/mongoRequestApprove.spec.ts
@@ -10,16 +10,20 @@ test.describe('Mongo Request Approval', () => {
         await loginAsMongoParent(page);
     });
 
+    test.afterEach(async ({request}) => {
+        if(currentTransactionId) {
+            //Cleanup
+            const deleted = await request.delete(`${await baseUrl()}/api/transactions/test/${currentTransactionId}`);
+            expect(deleted.ok(), 'should cleanup transaction properly').toBeTruthy();
+            expect(await deleted.json() as unknown as DeleteResponse, 'should expect 1 item deleted').toEqual({deletedCount: 1});
+        }
+    });
+
     test('should approve request', async ({page, request}) => {
         const requestComponent = page.getByText(`Request $10Reason: ${currentTransactionId}`);
         await expect(requestComponent).toBeVisible();
         await requestComponent.getByRole('button', {name: 'Approve'}).click();
         await expect(requestComponent).toBeHidden();
-
-        //Cleanup
-        const deleted = await request.delete(`${await baseUrl()}/api/transactions/test/${currentTransactionId}`);
-        expect(deleted.ok(), 'should cleanup transaction properly').toBeTruthy();
-        expect(await deleted.json() as unknown as DeleteResponse, 'should expect 1 item deleted').toEqual({deletedCount: 1});
     });
 
 });
